Changed xmega_hardware_interface.yaml and battery voltage from float to double

+++ pandora_xmega_hardware_interface/include/pandora_xmega_hardware_interface/battery_sensor.h
@@ -53,8 +53,8 @@ class BatterySensor : virtual public SensorBase
 
  public:
   
-  float psuVoltage;
-  float motorVoltage;
+  double psuVoltage;
+  double motorVoltage;
 };
 
 } // namespace pandora_xmega
